+++ zandbox/src/controller/contract/query/mod.rs
@@ -10,7 +10,6 @@ use std::sync::RwLock;
 
 use actix_web::http::StatusCode;
 use actix_web::web;
-use actix_web::Responder;
 use serde_json::Value as JsonValue;
 
 use zinc_build::Value as BuildValue;
@@ -22,104 +21,79 @@ use crate::response::Response;
 use crate::shared_data::SharedData;
 
 use self::error::Error;
-use self::request::Body;
-use self::request::Query;
+use self::request::Body as RequestBody;
+use self::request::Query as RequestQuery;
 
 ///
 /// The HTTP request handler.
 ///
 pub async fn handle(
     app_data: web::Data<Arc<RwLock<SharedData>>>,
-    query: web::Query<Query>,
-    body: web::Json<Body>,
-) -> impl Responder {
+    query: web::Query<RequestQuery>,
+    body: web::Json<RequestBody>,
+) -> crate::Result<JsonValue, Error> {
     let query = query.into_inner();
     let body = body.into_inner();
 
-    let contract = match app_data
+    let contract = app_data
         .read()
         .expect(zinc_const::panic::MULTI_THREADING)
         .contracts
         .get(&query.contract_id)
         .cloned()
-    {
-        Some(contract) => contract,
-        None => return Response::error(Error::ContractNotFound),
-    };
+        .ok_or(Error::ContractNotFound)?;
 
-    let storage_value = match app_data
+    let storage_value = app_data
         .read()
         .expect(zinc_const::panic::MULTI_THREADING)
         .postgresql_client
         .select_fields(FieldSelectInput::new(query.contract_id))
         .await
-    {
-        Ok(output) => {
-            if output.len() != contract.build.storage.len() {
-                return Response::error(Error::InvalidStorageSize {
-                    expected: contract.build.storage.len(),
-                    found: output.len(),
-                });
-            }
+        .map_err(Error::Database)?;
+    if storage_value.len() != contract.build.storage.len() {
+        return Err(Error::InvalidStorageSize {
+            expected: contract.build.storage.len(),
+            found: storage_value.len(),
+        });
+    }
+    let mut fields = Vec::with_capacity(storage_value.len());
+    for (index, FieldSelectOutput { name, value }) in storage_value.into_iter().enumerate() {
+        let r#type = contract.build.storage[index].1.clone();
+        let value = match BuildValue::try_from_typed_json(value, r#type) {
+            Ok(value) => value,
+            Err(error) => return Err(Error::InvalidStorage(error)),
+        };
+        fields.push((name, value))
+    }
+    let storage_value = BuildValue::Contract(fields);
 
-            let mut fields = Vec::with_capacity(output.len());
-            for (index, FieldSelectOutput { name, value }) in output.into_iter().enumerate() {
-                let r#type = contract.build.storage[index].1.clone();
-                let value = match BuildValue::try_from_typed_json(value, r#type) {
-                    Ok(value) => value,
-                    Err(error) => return Response::error(Error::InvalidStorage(error)),
-                };
-                fields.push((name, value))
-            }
-            BuildValue::Contract(fields)
-        }
-        Err(error) => return Response::error(Error::Database(error)),
+    let method_name = match query.method {
+        Some(method_name) => method_name,
+        None => return Response::new_with_data(StatusCode::OK, storage_value.into_json()).into(),
     };
 
-    match query.method {
-        Some(method_name) => {
-            let method = match contract.build.methods.get(method_name.as_str()).cloned() {
-                Some(method) => method,
-                None => return Response::error(Error::MethodNotFound),
-            };
-
-            let arguments = match body.arguments {
-                Some(arguments) => arguments,
-                None => return Response::error(Error::MethodArgumentsNotFound),
-            };
-
-            let input_value = match BuildValue::try_from_typed_json(arguments, method.input) {
-                Ok(input_value) => input_value,
-                Err(error) => return Response::error(Error::InvalidInput(error)),
-            };
-
-            if method.is_mutable {
-                return Response::error(Error::MethodIsMutable);
-            }
-
-            let output = match async_std::task::spawn_blocking(move || {
-                zinc_vm::ContractFacade::new(contract.build).run::<Bn256>(
-                    input_value,
-                    storage_value,
-                    method_name,
-                )
-            })
-            .await
-            {
-                Ok(output) => output,
-                Err(error) => return Response::error(Error::RuntimeError(error)),
-            };
-
-            dbg!(output.transfers);
-
-            Response::<JsonValue, Error>::success_with_data(
-                StatusCode::OK,
-                output.result.into_json(),
-            )
-        }
-        None => Response::<JsonValue, Error>::success_with_data(
-            StatusCode::OK,
-            storage_value.into_json(),
-        ),
+    let method = contract
+        .build
+        .methods
+        .get(method_name.as_str())
+        .cloned()
+        .ok_or(Error::MethodNotFound)?;
+    let arguments = body.arguments.ok_or(Error::MethodArgumentsNotFound)?;
+    let input_value =
+        BuildValue::try_from_typed_json(arguments, method.input).map_err(Error::InvalidInput)?;
+    if method.is_mutable {
+        return Err(Error::MethodIsMutable);
     }
+
+    let output = async_std::task::spawn_blocking(move || {
+        zinc_vm::ContractFacade::new(contract.build).run::<Bn256>(
+            input_value,
+            storage_value,
+            method_name,
+        )
+    })
+    .await
+    .map_err(Error::RuntimeError)?;
+
+    Response::new_with_data(StatusCode::OK, output.result.into_json()).into()
 }