+++ zinc-compiler/src/syntax/parser/statement/local_fn.rs
@@ -155,7 +155,6 @@ mod tests {
                 Location::new(1, 1),
                 ExpressionTreeNode::operand(ExpressionOperand::Block(BlockExpression::new(
                     Location::new(1, 1),
-                    false,
                     vec![],
                     Some(ExpressionTree::new(
                         Location::new(1, 3),